+++ indico/modules/events/reminders/forms.py
@@ -51,8 +51,9 @@ class ReminderForm(IndicoForm):
         self.event = kwargs.pop('event')
         self.timezone = self.event.timezone
         super().__init__(*args, **kwargs)
-        self.reply_to_address.choices = (list(self.event
-                                         .get_allowed_sender_emails(extra=self.reply_to_address.object_data).items()))
+        allowed_senders = self.event.get_allowed_sender_emails(include_noreply=True,
+                                                               extra=self.reply_to_address.object_data)
+        self.reply_to_address.choices = list(allowed_senders.items())
         if self.event.type_ == EventType.lecture:
             del self.include_summary